+++ Package.swift
@@ -4,43 +4,46 @@
 import PackageDescription
 
 let package = Package(
-    name: "EudiWalletKit",
+	name: "EudiWalletKit",
 	platforms: [.macOS(.v13), .iOS(.v15)],
-    products: [
-        // Products define the executables and libraries a package produces, making them visible to other packages.
-        .library(
-            name: "EudiWalletKit",
-            targets: ["EudiWalletKit"]),
-    ],
-    dependencies: [
-    .package(url: "https://github.com/apple/swift-log.git", from: "1.5.3"),
-		.package(url: "https://github.com/crspybits/swift-log-file", from: "0.1.0"),
-		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-iso18013-data-transfer.git",  exact: "0.5.5"),
-		//.package(path: "../eudi-lib-ios-iso18013-data-transfer"),
-		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-wallet-storage.git", exact: "0.4.5"),
-    .package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-sdjwt-swift.git", exact: "0.6.0"),
-		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-siop-openid4vp-swift.git", exact: "0.7.0"),
-//		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-openid4vci-swift.git", exact: "0.12.1"),
-	.package(url: "https://github.com/german-first-iteration/eudi-lib-ios-openid4vci-swift.git", branch: "eudi-lib-ios-openid4vci-swift-0.12.0"),
+	products: [
+		// Products define the executables and libraries a package produces, making them visible to other packages.
+		.library(
+			name: "EudiWalletKit",
+			targets: ["EudiWalletKit"])
 	],
-    targets: [
-        // Targets are the basic building blocks of a package, defining a module or a test suite.
-        // Targets can depend on other targets in this package and products from dependencies.
-        .target(
-            name: "EudiWalletKit", dependencies: [
-		    .product(name: "MdocDataTransfer18013", package: "eudi-lib-ios-iso18013-data-transfer"),
+	dependencies: [
+		.package(url: "https://github.com/apple/swift-log.git", from: "1.5.3"),
+		.package(url: "https://github.com/crspybits/swift-log-file", from: "0.1.0"),
+		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-iso18013-data-transfer.git",  exact: "0.6.0"),
+		// .package(path: "../eudi-lib-ios-iso18013-data-model"),
+		// .package(path: "../eudi-lib-ios-iso18013-data-transfer"),
+		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-wallet-storage.git", exact: "0.4.8"),
+		// .package(path: "../eudi-lib-ios-wallet-storage"),
+		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-sdjwt-swift.git",exact: "0.6.0"),
+		.package(url:"https://github.com/eu-digital-identity-wallet/eudi-lib-ios-siop-openid4vp-swift.git",exact: "0.8.1"),
+		.package(url: "https://github.com/eu-digital-identity-wallet/eudi-lib-ios-openid4vci-swift.git",exact: "0.12.1"),
+	],
+	targets: [
+		// Targets are the basic building blocks of a package, defining a module or a test suite.
+		// Targets can depend on other targets in this package and products from dependencies.
+		.target(
+			name: "EudiWalletKit",
+			dependencies: [
+				//.product(name: "MdocDataModel18013", package: "eudi-lib-ios-iso18013-data-model"),
+				.product(name: "MdocDataTransfer18013", package: "eudi-lib-ios-iso18013-data-transfer"),
 				.product(name: "WalletStorage", package: "eudi-lib-ios-wallet-storage"),
 				.product(name: "SiopOpenID4VP", package: "eudi-lib-ios-siop-openid4vp-swift"),
 				.product(name: "OpenID4VCI", package: "eudi-lib-ios-openid4vci-swift"),
-        .product(name: "eudi-lib-sdjwt-swift", package: "eudi-lib-sdjwt-swift"),
-        .product(name: "Logging", package: "swift-log"),
+				.product(name: "eudi-lib-sdjwt-swift", package: "eudi-lib-sdjwt-swift"),
+				.product(name: "Logging", package: "swift-log"),
 				.product(name: "FileLogging", package: "swift-log-file"),
-          ]
-        ),
-        .testTarget(
-            name: "EudiWalletKitTests",
-            dependencies: ["EudiWalletKit"],
-						resources: [.process("Resources")]
-						)
-    ]
+			]
+		),
+		.testTarget(
+			name: "EudiWalletKitTests",
+			dependencies: ["EudiWalletKit"],
+			resources: [.process("Resources")]
+		),
+	]
 )